+++ _modules/azure.js
@@ -5,11 +5,38 @@
 const echo  = require("./echo");
 const shell = require("shelljs");
 
+// #endregion Imports
+
+// -----------------------------------------------------------------------------------------
+// #region Variables
+// -----------------------------------------------------------------------------------------
+
+const pythonInstallerUrl  = "https://www.python.org/ftp/python/3.7.4/python-3.7.4-amd64.exe";
+
+// #endregion Variables
+
 // -----------------------------------------------------------------------------------------
 // #region Functions
 // -----------------------------------------------------------------------------------------
 
 const azure = {
+    validateAzCli() {
+        if (!shell.which("az")) {
+            echo.message("Azure CLI not found. Attempting install via PIP...");
+
+            if (!shell.which("pip")) {
+                echo.error(`PIP is required - ${pythonInstallerUrl}`);
+                shell.exit(1);
+            }
+
+            if (shell.exec("pip install azure-cli").code !== 0) {
+                echo.error("Failed to install azure cli via pip");
+                shell.exit(1);
+            }
+
+            echo.success(" - Successfully installed Azure CLI");
+        }
+    },
     /**
      * Executes `az login` using username/password.
      * @param {*} username Azure account username
@@ -17,9 +44,11 @@ const azure = {
      */
     login(username, secret) {
         const loginCommand = `az login -u ${username} -p ${secret}`;
-        if (shell.exec(loginCommand).code !== 0) {
+
+        const cmdResult = shell.exec(loginCommand);
+        if (cmdResult.code !== 0) {
             echo.error(" - Failed to login to Azure");
-            shell.exit(1);
+            shell.exit(cmdResult);
         }
     },
     /**
@@ -30,9 +59,11 @@ const azure = {
      */
     login(clientId, tenantId, secret) {
         const loginCommand = `az login --service-principal -u ${clientId} -t ${tenantId} -p=${secret}`;
-        if (shell.exec(loginCommand).code !== 0) {
+
+        const cmdResult = shell.exec(loginCommand);
+        if (cmdResult.code !== 0) {
             echo.error(" - Failed to login to Azure");
-            shell.exit(1);
+            shell.exit(cmdResult.code);
         }
     },
     /**
@@ -40,9 +71,11 @@ const azure = {
      */
     logout() {
         const logoutCmd = "az logout";
-        if (shell.exec(logoutCmd).code !== 0) {
+
+        const cmdResult = shell.exec(logoutCmd);
+        if (cmdResult.code !== 0) {
             echo.error(" - Failed to logout from Azure, it is recommended `az logout` is run manually");
-            shell.exit(1);
+            shell.exit(cmdResult.code);
         }
     },
 }
