+++ _modules/deploy-config.js
@@ -1,0 +1,54 @@
+// -----------------------------------------------------------------------------------------
+// #region Imports
+// -----------------------------------------------------------------------------------------
+
+const echo         = require("./echo");
+const file         = require("./file");
+const frontendPath = require("./frontend-path");
+const path         = require("path");
+const shell        = require("shelljs");
+const upath        = require("upath");
+
+// #endregion Imports
+
+// -----------------------------------------------------------------------------------------
+// #region Variables
+// -----------------------------------------------------------------------------------------
+
+const environmentFileName = ".env.local";
+
+// #endregion Variables
+
+// -----------------------------------------------------------------------------------------
+// #region Functions
+// -----------------------------------------------------------------------------------------
+
+const deployConfig = {
+    /**
+     * Generates a local env file with a PUBLIC_URL variable.
+     * @param {*} publicUrl the value to set PUBLIC_URL to.
+     */
+    configurePublicUrl(publicUrl) {
+        // Delete current file if it exists
+        const environmentFilePath = upath.toUnix(path.join(frontendPath.projectDir(), environmentFileName));
+        file.deleteIfExists(environmentFilePath);
+
+        // Output environment file configured with 'PUBLIC_URL'
+        echo.message(`Generating new ${environmentFilePath} file with contents of...`);
+        const publicUrlLine = `PUBLIC_URL=${publicUrl}`;
+        shell.echo("# File generated by `and-cli`").toEnd(environmentFilePath);
+        shell.echo(publicUrlLine).toEnd(environmentFilePath);
+        shell.echo("").toEnd(environmentFilePath);
+        echo.success("Local environment file generated successfully");
+    },
+}
+
+// #endregion Functions
+
+// -----------------------------------------------------------------------------------------
+// #region Exports
+// -----------------------------------------------------------------------------------------
+
+module.exports = deployConfig;
+
+// #endregion Exports
