+++ and-cli-deploy-aws-s3.js
@@ -4,13 +4,11 @@ require("./command-runner").run(async () => {
     // #region Imports
     // -----------------------------------------------------------------------------------------
 
+    const deployConfig   = require("./_modules/deploy-config");
     const echo           = require("./_modules/echo");
-    const file           = require("./_modules/file");
     const frontendPath   = require("./_modules/frontend-path");
-    const path           = require("path");
     const program        = require("commander");
     const shell          = require("shelljs");
-    const upath          = require("upath");
     const webpackPublish = require("./_modules/webpack-publish");
 
     // #endregion Imports
@@ -19,7 +17,6 @@ require("./command-runner").run(async () => {
     // #region Variables
     // -----------------------------------------------------------------------------------------
 
-    const environmentFileName = ".env.local";
     let   destination         = null;
     let   profile             = null;
     const pythonInstallerUrl  = "https://www.python.org/ftp/python/3.7.4/python-3.7.4-amd64.exe";
@@ -47,17 +44,7 @@ require("./command-runner").run(async () => {
 
             // Configure .env.local to set public url before publish
             if (program.publicUrl) {
-                // Delete current file if it exists
-                const environmentFilePath = upath.toUnix(path.join(frontendPath.projectDir(), environmentFileName));
-                file.deleteIfExists(environmentFilePath);
-
-                // Output environment file configured with 'PUBLIC_URL'
-                echo.message(`Generating new ${environmentFilePath} file with contents of...`);
-                const publicUrlLine = `PUBLIC_URL=${program.publicUrl}`;
-                shell.echo("# File is generated by the `and-cli deploy aws-s3` command").toEnd(environmentFilePath);
-                shell.echo(publicUrlLine).toEnd(environmentFilePath);
-                shell.echo("").toEnd(environmentFilePath);
-                echo.success("Local environment file generated successfully");
+                deployConfig.configurePublicUrl(program.publicUrl);
             }
 
             // Locally publish frontend via webpack
