+++ and-cli-deploy-azure-storage.js
@@ -5,13 +5,11 @@ require("./command-runner").run(async () => {
     // -----------------------------------------------------------------------------------------
 
     const azure          = require("./_modules/azure");
+    const deployConfig   = require("./_modules/deploy-config");
     const echo           = require("./_modules/echo");
-    const file           = require("./_modules/file");
     const frontendPath   = require("./_modules/frontend-path");
-    const path           = require("path");
     const program        = require("commander");
     const shell          = require("shelljs");
-    const upath          = require("upath");
     const webpackPublish = require("./_modules/webpack-publish");
 
     // #endregion Imports
@@ -22,7 +20,6 @@ require("./command-runner").run(async () => {
 
     let   clientId            = null;
     let   destination         = null;
-    const environmentFileName = ".env.local";
     const pythonInstallerUrl  = "https://www.python.org/ftp/python/3.7.4/python-3.7.4-amd64.exe";
     let   recursive           = false;
     let   secret              = null;
@@ -55,17 +52,7 @@ require("./command-runner").run(async () => {
 
             // Configure .env.local to set public url before publish
             if (program.publicUrl) {
-                // Delete current file if it exists
-                const environmentFilePath = upath.toUnix(path.join(frontendPath.projectDir(), environmentFileName));
-                file.deleteIfExists(environmentFilePath);
-
-                // Output environment file configured with 'PUBLIC_URL'
-                echo.message(`Generating new ${environmentFilePath} file with contents of...`);
-                const publicUrlLine = `PUBLIC_URL=${program.publicUrl}`;
-                shell.echo("# File is generated by the `and-cli deploy azure-storage` command").toEnd(environmentFilePath);
-                shell.echo(publicUrlLine).toEnd(environmentFilePath);
-                shell.echo("").toEnd(environmentFilePath);
-                echo.success("Local environment file generated successfully");
+                deployConfig.configurePublicUrl(program.publicUrl);
             }
 
             // Locally publish frontend via webpack
